add hapi 17 compatible module
Users would `require('circonus-cir/hapi')` to use this module.
The reason is that hapi 17 uses backwards incompatible syntax,
so users on Node < 8 would not even be able to import the
module if it was exposed via index.js. Users of hapi < 17
can require the legacy module or get it via index.js.